requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
astunparse==1.6.3
blinker==1.9.0
blis==1.0.2
catalogue==2.0.10
certifi @ file:///C:/b/abs_1fw_exq1si/croot/certifi_1725551736618/work/certifi
cffi @ file:///D:/bld/cffi_1725560649097/work
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
contourpy @ file:///D:/bld/contourpy_1731428338157/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1733332471406/work
cymem==2.0.10
filelock==3.16.1
Flask==3.1.0
flatbuffers==24.3.25
fonttools @ file:///D:/bld/fonttools_1733908982599/work
future @ file:///home/conda/feedstock_root/build_artifacts/future_1733754212340/work
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.1
h5py==3.12.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
keras==3.7.0
kiwisolver @ file:///D:/bld/kiwisolver_1725459382062/work
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
ml-dtypes==0.4.1
monai @ file:///home/conda/feedstock_root/build_artifacts/monai_1729125401112/work
mpmath==1.3.0
munkres==1.1.4
murmurhash==1.0.11
namex==0.0.8
networkx==3.4.2
numpy @ file:///C:/b/abs_c1ywpu18ar/croot/numpy_and_numpy_base_1708638681471/work/dist/numpy-1.26.4-cp310-cp310-win_amd64.whl#sha256=ebb5aa2b36d8afa5ec3231c19e5a1fc75b6d85e7db483f0fb9e77dad58469977
opt_einsum==3.4.0
optree==0.13.1
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
pandas==2.2.3
pillow @ file:///D:/bld/pillow_1729065755620/work
preshed==3.0.9
protobuf==5.29.1
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic==2.10.3
pydantic_core==2.27.1
pydicom @ file:///home/conda/feedstock_root/build_artifacts/pydicom_1734037635475/work
Pygments==2.18.0
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1733222594562/work
PySide6==6.8.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
pytz==2024.2
PyYAML @ file:///D:/bld/pyyaml_1725456311802/work
requests==2.32.3
rich==13.9.4
scikit-learn==1.5.2
scipy==1.14.1
shellingham==1.5.4
shiboken6==6.8.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
smart-open==7.0.5
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.0
sympy==1.13.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
thinc==8.3.2
threadpoolctl==3.5.0
torch==2.0.1+cu118
tornado @ file:///D:/bld/tornado_1732615925919/work
tqdm==4.67.1
typer==0.15.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1733188668063/work
tzdata==2024.2
unicodedata2 @ file:///D:/bld/unicodedata2_1729704580516/work
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.0
